refactor out a consensus data fetcher from table router
