+++ network/src/tests/mod.rs
@@ -17,11 +17,11 @@
 //! Tests for polkadot and consensus network.
 
 use super::{PolkadotProtocol, Status, Message, FullStatus};
-use consensus::{CurrentConsensus, Knowledge};
+use consensus::{CurrentConsensus, ConsensusParams, Knowledge};
 
 use parking_lot::Mutex;
 use polkadot_consensus::GenericStatement;
-use polkadot_primitives::{Block, SessionKey};
+use polkadot_primitives::{Block, Hash, SessionKey};
 use polkadot_primitives::parachain::{CandidateReceipt, HeadData, BlockData};
 use substrate_primitives::H512;
 use codec::Encode;
@@ -87,9 +87,12 @@ fn make_status(status: &Status, roles: Roles) -> FullStatus {
 	}
 }
 
-fn make_consensus(local_key: SessionKey) -> (CurrentConsensus, Arc<Mutex<Knowledge>>) {
-	let knowledge = Arc::new(Mutex::new(Knowledge::new()));
-	let c = CurrentConsensus::new(knowledge.clone(), local_key);
+fn make_consensus(parent_hash: Hash, local_key: SessionKey) -> (CurrentConsensus, Arc<Mutex<Knowledge>>) {
+	let c = CurrentConsensus::new(ConsensusParams {
+		local_session_key: Some(local_key),
+		parent_hash,
+	});
+	let knowledge = c.knowledge().clone();
 
 	(c, knowledge)
 }
@@ -119,8 +122,8 @@ fn sends_session_key() {
 
 	{
 		let mut ctx = TestContext::default();
-		let (consensus, _knowledge) = make_consensus(local_key);
-		protocol.new_consensus(&mut ctx, parent_hash, consensus);
+		let (consensus, _knowledge) = make_consensus(parent_hash, local_key);
+		protocol.new_consensus(&mut ctx, consensus);
 		assert!(ctx.has_message(peer_a, Message::SessionKey(local_key)));
 	}
 
@@ -159,8 +162,8 @@ fn fetches_from_those_with_knowledge() {
 
 	let status = Status { collating_for: None };
 
-	let (consensus, knowledge) = make_consensus(local_key);
-	protocol.new_consensus(&mut TestContext::default(), parent_hash, consensus);
+	let (consensus, knowledge) = make_consensus(parent_hash, local_key);
+	protocol.new_consensus(&mut TestContext::default(), consensus);
 
 	knowledge.lock().note_statement(a_key, &GenericStatement::Valid(candidate_hash));
 	let recv = protocol.fetch_block_data(&mut TestContext::default(), &candidate_receipt, parent_hash);
@@ -289,11 +292,11 @@ fn many_session_keys() {
 	let local_key_a = [3; 32].into();
 	let local_key_b = [4; 32].into();
 
-	let (consensus_a, _knowledge_a) = make_consensus(local_key_a);
-	let (consensus_b, _knowledge_b) = make_consensus(local_key_b);
+	let (consensus_a, _knowledge_a) = make_consensus(parent_a, local_key_a);
+	let (consensus_b, _knowledge_b) = make_consensus(parent_b, local_key_b);
 
-	protocol.new_consensus(&mut TestContext::default(), parent_a, consensus_a);
-	protocol.new_consensus(&mut TestContext::default(), parent_b, consensus_b);
+	protocol.new_consensus(&mut TestContext::default(), consensus_a);
+	protocol.new_consensus(&mut TestContext::default(), consensus_b);
 
 	assert_eq!(protocol.live_consensus.recent_keys(), &[local_key_a, local_key_b]);
